+++ backend/bin/run-jobs.sh
@@ -8,6 +8,7 @@ DIR=$(dirname "$(realpath "$0")")
 "${DIR}"/console check-tokens --log --hide-details
 "${DIR}"/console update-member-tracking --log
 "${DIR}"/console auto-allowlist --log --hide-details
+"${DIR}"/console update-service-accounts --log --hide-details
 "${DIR}"/console send-invalid-token-mail --log
 "${DIR}"/console send-missing-character-mail --log
 "${DIR}"/console clean-http-cache --log