+++ backend/tests/Functional/Controller/User/ServiceControllerTest_TestService.php
@@ -62,4 +62,9 @@ class ServiceControllerTest_TestService implements ServiceInterface
             return 'new-pass';
         }
     }
+
+    public function getAllAccounts(): array
+    {
+        throw new Exception();
+    }
 }